Updated launchd plist with KeepAlive dict.
This plist is compatible with 10.5 and newer (sorry for 10.4…).
Now, when we stop CentralReport with the command line, CentralReport send the "0" exit code. And OS X will not restart CR anymore.
But, when CR not send 0 exit code (bug, error, or any else), OS X try to restart it automatically.