Improve installation scripts and new startup script for debian init.d
Bash scripts are stored in "bash" folder, introducing functions use.
Updated install and uninstall scripts to working with the new functions.

Added : centralreport_debian.sh, will be useful to startup CentralReport when Debian is starting. This script will be go to "/etc/init.d", when the installer will be updated, in a next commit.