Configuration editor and configuration class updated.
You can execute "python config.py" to edit CentralReport configuration. The script edit the config file (/etc/cr/centralreport.cfg), asking some questions to the user.
The script can only enable or disable the internal web server, and change the HTTP port. The script check if the new port is free or not.

In a next commit, this script will be called from the install.sh.