Added: static files and 'data' module.
We have a new module, called 'data'. It contains data about the current host and the daemon. For now, just used to store the last check (cr.entities.checks.Check) and the host_info object (cr.entities.host.Info).
The static files can contains… static files, used for the web services, like the JSON templates for the exchanges with the Online part.